Add files via upload
Effective parking solutions are essential in the fast-paced world of urban growth. This need is met by our initiative, "Smart Parking," which makes use of advanced object detection technology. We use YOLO V8, a state-of-the-art algorithm that is renowned for its quickness and precision in object recognition in photos. This project combines a number of Python packages, including cv2 for image processing tasks, OpenCV, Pandas for data management, and NumPy for numerical operations.
Our technology uses a network of cameras that are positioned thoughtfully all across a parking lot to record visuals in real time. After that, YOLO V8 processes these pictures along with OpenCV and cv2 to identify open parking spots. While Pandas organize the data into a logical style to enable efficient space identification, NumPy speeds up the computing process.
Our Smart Parking system's installation should result in a notable decrease in the amount of time spent looking for parking spaces. This helps reduce traffic in parking lots. Because of the system's adaptability, it can be used in a variety of settings, including airports and shopping malls. All things considered; our project serves as an example of how artificial intelligence may be incorporated into urban infrastructure to improve everyday living.​